requirements.txt
Pillow
tqdm
tk
pytz
pyinstaller
